+++ Realm/RLMResults_Private.hpp
@@ -38,7 +38,10 @@ NS_ASSUME_NONNULL_BEGIN
  */
 - (instancetype)initWithResults:(realm::Results)results;
 
-+ (instancetype)resultsWithObjectInfo:(RLMClassInfo&)info results:(realm::Results)results;
+- (instancetype)initWithObjectInfo:(RLMClassInfo&)info results:(realm::Results&&)results;
++ (instancetype)resultsWithObjectInfo:(RLMClassInfo&)info results:(realm::Results&&)results;
+
+- (instancetype)subresultsWithResults:(realm::Results)results;
 @end
 
 NS_ASSUME_NONNULL_END